refactor: organize E2E tests and add LLM provider tests
- Move deterministic tests to tests/e2e/deterministic/
- Add tests/e2e/llm/ with real provider tests (anthropic, openai, google)
- Add LLM models to test server (haiku, gpt-4.1-nano, gemini-2.0-flash)
- Replace Playwright projects with env var config (E2E_TEST_DIR, E2E_VIDEO)
- Remove placeholder e2e-llm workflow, add LLM step to main CI
- Remove committed __pycache__, add to .gitignore
- Document test commands in CLAUDE.md
