+++ cspell.json
@@ -1,5 +1,5 @@
 {
   "version": "0.2",
   "language": "en",
-  "words": ["oxlint", "Pydantic", "pydantic", "regen", "shiki", "uidotdev", "usehooks", "uvicorn"]
+  "words": ["logfire", "Pydantic", "pydantic", "uvicorn"]
 }
