+++ tests/e2e/deterministic/multi-tool.spec.ts
@@ -1,6 +1,6 @@
 import { test, expect } from '@playwright/test'
-import { sendMessage } from './conversation'
-import { toolCard } from './tools'
+import { sendMessage } from '../conversation'
+import { toolCard } from '../tools'
 
 test.describe('multi-tool', () => {
   test('shows completed status for both tools', async ({ page }) => {
