+++ tests/e2e/deterministic/tool-approval.spec.ts
@@ -1,6 +1,6 @@
 import { test, expect } from '@playwright/test'
-import { sendMessage } from './conversation'
-import { toolCard } from './tools'
+import { sendMessage } from '../conversation'
+import { toolCard } from '../tools'
 
 test.describe('tool approval', () => {
   test('tool approval accepted', async ({ page }) => {
