+++ tests/e2e/deterministic/tool-calls.spec.ts
@@ -1,6 +1,6 @@
 import { test, expect } from '@playwright/test'
-import { sendMessage } from './conversation'
-import { toolCard } from './tools'
+import { sendMessage } from '../conversation'
+import { toolCard } from '../tools'
 
 test.describe('tool calls', () => {
   test('shows tool call UI with tool name', async ({ page }) => {
